This program uses while and if/else.
I only study if/else in college, but I try to update my algorithms myself. The simple structure without the repeating loops would be:
#include <iostream>
using namespace std;
int main(){
	char a;
	cout << "Olá... que turno você estuda? M - Matutino | V - Verspertino | N - Noturno" << endl;
	cin >> a;
	if (a == 'm'){
		cout << "Bom dia!\n";
	} else if (a == 'v'){
		cout <<"Boa tarde!\n";
	} else if (a == 'n'){
		cout << "Boa noite!\n";
	} else {
	cout <<	"Valor inválido.\n";
	}
system("pause");
 return 0;
}

Important, it would be possible to do this algorithm with a string, but I would change the form of 'a' to "a" and I will still try to improve the algorithm for the case sensitive function.